backend/retrieval_pipeline/pyproject.toml
[project]
name = "rag-retrieval-pipeline"
version = "0.1.0"
description = "RAG Retrieval Pipeline for Physical AI Textbook"
authors = [{name = "Developer", email = "dev@example.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "qdrant-client>=1.3.0",
    "cohere>=4.0.0",
    "python-dotenv>=0.19.0",
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
]